Political-Chrome-Extension/HttpTrigger/requirements.txt
azure-functions
azure-ai-textanalytics
newspaper3k